restructure format | #16
