--- api_files/routers/data/group_scripts/_router_groups.py
@@ -1,8 +1,0 @@
-from fastapi import APIRouter
-from api_files.routers.data.group_scripts import return_groups
-router = APIRouter(
-    prefix="/groups",
-    tags=['Card Groups']
-)
-
-router.include_router(return_groups.router)
