+++ preordain/information/router.py
@@ -1,44 +1,23 @@
-from fastapi import APIRouter, Response, status
-from api_files.models import BaseResponse, CardInformation
-# from api_files.response_class.pretty import PrettyJSONResp
-from psycopg.rows import dict_row
-import scripts.connect.to_database as to_db
+import logging
+log = logging.getLogger()
+from preordain.utils.connections import connect_db, send_response
+from fastapi import APIRouter, Response, status, Depends
+from fastapi.exceptions import HTTPException
+from preordain.information.utils import parse_data_for_response
+from preordain.information.models import CardInformation
+from preordain.models import BaseResponse
+from preordain.dependencies import write_access
 
 
-def parse_data_for_response(data: list):
-    """
-    Parse the data you recieved for this format.
-    """
-    card_data = []
-    for cards in data:
-        card_data.append(
-            {
-                'name': cards['name'],
-                'set': cards['set'],
-                'set_full': cards['set_full'],
-                'id': cards['id'],
-                'last_updated': cards['last_updated'],
-                'prices': {
-                    'usd': cards['usd'],
-                    'usd_foil': cards['usd_foil'],
-                    'euro': cards['euro'],
-                    'euro_foil': cards['euro_foil'],
-                    'tix': cards['tix'],
-                }
-            }
-        )
-    return card_data
-
-router = APIRouter(
-    prefix="/search",
-)
+user_router = APIRouter(prefix="/search",tags=["Card Information, etc"])
+admin_router = APIRouter(prefix="/admin", tags=["Admin Panel"], dependencies=[Depends(write_access)])
 
 # Return all cards
-@router.get("/", 
+@user_router.get("/", 
     response_model=BaseResponse[CardInformation], 
-    description="Return all cards that are being tracked.",)
+    description="Return all cards that are being tracked.")
 async def read_items(response: Response):
-    conn, cur = to_db.connect_db(row_factory = dict_row)
+    conn, cur = connect_db()
     cur.execute("""
         SELECT 
             info.name,
@@ -71,18 +50,18 @@ async def read_items(response: Response):
         """
         )
     data = cur.fetchall()
+    conn.close()
     if data:
         response.status_code = status.HTTP_200_OK
         return BaseResponse(resp='card_info', status=response.status_code, data=parse_data_for_response(data))
     response.status_code = status.HTTP_404_NOT_FOUND
     return BaseResponse(resp='error_request', status=response.status_code, info={'message': 'There are no cards in the database!'})
 
-# Get a single card and return the data.
-@router.get("/{set}/{col_num}",
+@user_router.get("/{set}/{col_num}",
     description="Look for a specific card based on the set and collector number",
     response_model=BaseResponse[CardInformation])
 async def search_by_set_collector_num(set: str, col_num: str, response: Response):
-    conn, cur = to_db.connect_db(row_factory = dict_row)
+    conn, cur = connect_db()
     cur.execute(""" 
         SELECT 
             info.name,
@@ -127,12 +106,11 @@ async def search_by_set_collector_num(set: str, col_num: str, response: Response
     response.status_code = status.HTTP_404_NOT_FOUND
     return BaseResponse(resp='error_request', status=response.status_code)
 
-# # Filter for cards by their grouping.
-@router.get("/{group}",
+@user_router.get("/{group}",
     response_model=BaseResponse[CardInformation],
     description="Filter for cards by their groups.")
 async def find_by_group(group: str, response: Response):
-    conn, cur = to_db.connect_db(row_factory = dict_row)
+    conn, cur = connect_db()
     cur.execute(""" 
         
         SELECT   
@@ -181,3 +159,71 @@ async def find_by_group(group: str, response: Response):
         return BaseResponse[CardInformation](info={'group': group, 'info': 'ye'}, data=parse_data_for_response(data), resp='card_info', status=response.status_code)
     response.status_code = status.HTTP_404_NOT_FOUND
     return BaseResponse(resp='error_request', status=response.status_code)
+
+
+@admin_router.post("/add/{set}/{coll_num}")
+async def add_card_to_track(set: str, coll_num:str):
+
+    resp = send_response('GET',f"https://api.scryfall.com/cards/search?q=set:{set}+cn:{coll_num}")
+
+    try:
+        if resp['object'] != "list":
+            log.error("Not a card!")
+            raise HTTPException(
+                status_code=404, detail="This is not a card!"
+            )
+
+    except KeyError as e:
+        # ? What does this look like, again?
+        log.error(f"KeyError:{e}")
+
+    else:
+        if resp['total_cards'] != 1:
+            error_msg = f"Recieved list with more than 1. Set:{set}, ID:{coll_num}"
+            log.error(error_msg)
+            return error_msg
+
+        resp = resp['data'][0]
+        conn, cur = connect_db()
+        cur.execute("SELECT * from card_info.info where id = %s AND set= %s", (resp['collector_number'], resp['set']))
+        if not cur.fetchall(): # * Run this section if no results (empty lists are False)                    
+                if 'tcgplayer_etched_id' in resp:
+                    tcg_etched_id = resp['tcgplayer_etched_id']
+                else:
+                    tcg_etched_id = None
+
+                add_info_to_postgres = """
+                    INSERT INTO card_info.info (name, set, id, uri, tcg_id, tcg_id_etch, new_search)
+
+                    VALUES (%s,%s,%s,%s,%s,%s,%s)
+                    """
+                # ? Uncomment below in production.
+                cur.execute(add_info_to_postgres, (resp['name'], resp['set'], resp['collector_number'], resp['id'], resp['tcgplayer_id'], tcg_etched_id, True))
+                conn.commit()
+
+                log.info(f'Now tracking: {resp["name"]} from {resp["set_name"]}')
+                return f'Now tracking: {resp["name"]} from {resp["set_name"]}'
+
+        else:
+            log.info(f'Already tracking: {resp["name"]} from {resp["set_name"]}')
+            return f'Already tracking: {resp["name"]} from {resp["set_name"]}'
+
+@admin_router.delete("/remove/{set}/{coll_num}")
+async def remove_card_from_database(set:str, coll_num:str):
+
+    conn, cur = connect_db()
+
+    cur.execute("SELECT name, id, set from card_info.info where id = %s AND set = %s", (coll_num, set))
+
+    fetched_card = cur.fetchone()
+    if fetched_card:
+        text_resp = f"Deleting: {fetched_card['name']} (Set: {fetched_card['set']}, Collector Num: {fetched_card['id']})"
+        log.info(text_resp)
+        cur.execute("DELETE FROM card_info.info WHERE id = %s AND set = %s", (coll_num,set))
+        conn.commit()
+
+    else:
+        text_resp = f"Failed to delete, does not exist on DB (Set: {set}, Collector Num: {coll_num})"
+        log.error(text_resp)
+
+    return text_resp
