+++ preordain/utils/validators.py
@@ -1,0 +1,13 @@
+# from preordain.models import CardConditions, CardVariants
+
+# def verify_cond(condition: str):
+#     condition = condition.upper()
+#     if condition not in [cond for cond in CardConditions]:
+#         raise ValueError(f'Condition "{condition}" not a valid condition')
+#     return condition
+
+# def verify_vari(variant: str):
+#     variant = variant.title()
+#     if variant not in [var for var in CardVariants]:
+#         raise ValueError(f'Variant "{variant}" is not a valid variant')
+#     return variant
