+++ tests/test_db_connection.py
@@ -1,5 +1,5 @@
 import psycopg
-from scripts.config_reader import read_config
+from preordain.config_reader import read_config
 # TODO: Eventually
 
 db_params = read_config('CONNECT', 'database')
