+++ tests/test_scripts.py
@@ -1,6 +1,6 @@
 from scripts.check_if_update import check_date_to_update
 from scripts.update_config import update_config
-from scripts.config_reader import read_config
+from preordain.config_reader import read_config
 import pytest
 from datetime import timedelta
